code/requirements.txt
numpy
pandas
SQLAlchemy
openpyxl
#psycopg2 
psycopg2-binary #(Work for linux) 